0x0B-malloc_free

+++ 0x0B-malloc_free/2-str_concat.c
@@ -23,7 +23,7 @@ char *str_concat(char *s1, char *s2)
 		return (NULL);
 
 	for (i = 0; i < (long int) size; i++)
-		if (i < strlen(s1))
+		if (i < (long int) strlen(s1))
 			s[i] = s1[i];
 		else
 			s[i] = s2[i - strlen(s1)];
